backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
langchain==0.1.9
langchain-community==0.0.24
google-generativeai==0.3.2